+++ src/swap/central/godex.ts
@@ -22,16 +22,22 @@ import {
 } from 'edge-core-js/types'
 
 import {
-  checkInvalidCodes,
+  checkInvalidTokenIds,
   checkWhitelistedMainnetCodes,
   CurrencyPluginIdSwapChainCodeMap,
   getCodesWithTranscription,
   getMaxSwappable,
-  InvalidCurrencyCodes,
+  InvalidTokenIds,
   makeSwapPluginQuote,
   SwapOrder
 } from '../../util/swapHelpers'
-import { convertRequest, getAddress, memoType } from '../../util/utils'
+import {
+  convertRequest,
+  denominationToNative,
+  getAddress,
+  memoType,
+  nativeToDenomination
+} from '../../util/utils'
 import { asNumberString, EdgeSwapRequestPlugin } from '../types'
 
 const pluginId = 'godex'
@@ -82,14 +88,20 @@ const asQuoteInfo = asObject({
   return_extra_id: asEither(asString, asNull)
 })
 
-const INVALID_CURRENCY_CODES: InvalidCurrencyCodes = {
+const INVALID_TOKEN_IDS: InvalidTokenIds = {
   from: {
     digibyte: 'allCodes',
-    polygon: ['USDC', 'USDC.e']
+    polygon: [
+      '3c499c542cef5e3811e1192ce70d8cc03d5c3359' /* USDC */,
+      '2791bca1f2de4661ed88a30c99a7a9449aa84174' /* USDC.e */
+    ]
   },
   to: {
-    polygon: ['USDC', 'USDC.e'],
-    zcash: ['ZEC'] // Godex doesn't support sending to unified addresses
+    polygon: [
+      '3c499c542cef5e3811e1192ce70d8cc03d5c3359' /* USDC */,
+      '2791bca1f2de4661ed88a30c99a7a9449aa84174' /* USDC.e */
+    ],
+    zcash: [null] // Godex doesn't support sending to unified addresses
   }
 }
 
@@ -208,13 +220,15 @@ export function makeGodexPlugin(opts: EdgeCorePluginOptions): EdgeSwapPlugin {
     // Convert the native amount to a denomination:
     const quoteAmount =
       request.quoteFor === 'from'
-        ? await request.fromWallet.nativeToDenomination(
+        ? nativeToDenomination(
+            request.fromWallet,
             request.nativeAmount,
-            request.fromCurrencyCode
+            request.fromTokenId
           )
-        : await request.toWallet.nativeToDenomination(
+        : nativeToDenomination(
+            request.toWallet,
             request.nativeAmount,
-            request.toCurrencyCode
+            request.toTokenId
           )
 
     // Swap the currencies if we need a reverse quote:
@@ -243,13 +257,15 @@ export function makeGodexPlugin(opts: EdgeCorePluginOptions): EdgeSwapPlugin {
     // min_amount returned could be for a different network than the user is requesting.
     // Check the minimum:
     const nativeMin = reverseQuote
-      ? await request.toWallet.denominationToNative(
+      ? denominationToNative(
+          request.toWallet,
           reply.min_amount,
-          request.toCurrencyCode
+          request.toTokenId
         )
-      : await request.fromWallet.denominationToNative(
+      : denominationToNative(
+          request.fromWallet,
           reply.min_amount,
-          request.fromCurrencyCode
+          request.fromTokenId
         )
 
     if (lt(request.nativeAmount, nativeMin)) {
@@ -298,16 +314,18 @@ export function makeGodexPlugin(opts: EdgeCorePluginOptions): EdgeSwapPlugin {
     log('sendReply' + JSON.stringify(sendReply, null, 2))
     const quoteInfo = asQuoteInfo(sendReply)
     const fromNativeAmount = floor(
-      await request.fromWallet.denominationToNative(
+      denominationToNative(
+        request.fromWallet,
         quoteInfo.deposit_amount,
-        request.fromCurrencyCode
+        request.fromTokenId
       ),
       0
     )
     const toNativeAmount = floor(
-      await request.toWallet.denominationToNative(
+      denominationToNative(
+        request.toWallet,
         quoteInfo.withdrawal_amount,
-        request.toCurrencyCode
+        request.toTokenId
       ),
       0
     )
@@ -380,7 +398,7 @@ export function makeGodexPlugin(opts: EdgeCorePluginOptions): EdgeSwapPlugin {
       opts: { promoCode?: string }
     ): Promise<EdgeSwapQuote> {
       const request = convertRequest(req)
-      checkInvalidCodes(INVALID_CURRENCY_CODES, request, swapInfo)
+      checkInvalidTokenIds(INVALID_TOKEN_IDS, request, swapInfo)
       checkWhitelistedMainnetCodes(
         MAINNET_CODE_TRANSCRIPTION,
         request,