+++ src/swap/central/swapuz.ts
@@ -22,18 +22,24 @@ import {
 
 import { div18 } from '../../util/biggystringplus'
 import {
-  checkInvalidCodes,
+  checkInvalidTokenIds,
   checkWhitelistedMainnetCodes,
   CurrencyPluginIdSwapChainCodeMap,
   ensureInFuture,
   getCodesWithTranscription,
   getMaxSwappable,
-  InvalidCurrencyCodes,
+  InvalidTokenIds,
   isLikeKind,
   makeSwapPluginQuote,
   SwapOrder
 } from '../../util/swapHelpers'
-import { convertRequest, getAddress, memoType } from '../../util/utils'
+import {
+  convertRequest,
+  denominationToNative,
+  getAddress,
+  memoType,
+  nativeToDenomination
+} from '../../util/utils'
 import { EdgeSwapRequestPlugin } from '../types'
 
 const pluginId = 'swapuz'
@@ -52,16 +58,22 @@ const asInitOptions = asObject({
 const orderUri = 'https://swapuz.com/order/'
 const uri = 'https://api.swapuz.com/api/home/v1/'
 
-const INVALID_CURRENCY_CODES: InvalidCurrencyCodes = {
+const INVALID_TOKEN_IDS: InvalidTokenIds = {
   from: {
-    ethereum: ['MATH'],
-    optimism: ['VELO'],
-    polygon: ['USDC', 'USDC.e']
+    ethereum: ['08d967bb0134f2d07f7cfb6e246680c53927dd30' /* MATH */],
+    optimism: ['9560e827af36c94d2ac33a39bce1fe78631088db' /* VELO */],
+    polygon: [
+      '3c499c542cef5e3811e1192ce70d8cc03d5c3359' /* USDC */,
+      '2791bca1f2de4661ed88a30c99a7a9449aa84174' /* USDC.e */
+    ]
   },
   to: {
-    ethereum: ['MATH'],
-    polygon: ['USDC', 'USDC.e'],
-    zcash: ['ZEC'],
+    ethereum: ['08d967bb0134f2d07f7cfb6e246680c53927dd30' /* MATH */],
+    polygon: [
+      '3c499c542cef5e3811e1192ce70d8cc03d5c3359' /* USDC */,
+      '2791bca1f2de4661ed88a30c99a7a9449aa84174' /* USDC.e */
+    ],
+    zcash: [null],
     zksync: 'allCodes'
   }
 }
@@ -150,7 +162,7 @@ export function makeSwapuzPlugin(opts: EdgeCorePluginOptions): EdgeSwapPlugin {
   ): Promise<SwapOrder> => {
     const { fromWallet, toWallet } = request
 
-    checkInvalidCodes(INVALID_CURRENCY_CODES, request, swapInfo)
+    checkInvalidTokenIds(INVALID_TOKEN_IDS, request, swapInfo)
     checkWhitelistedMainnetCodes(MAINNET_CODE_TRANSCRIPTION, request, swapInfo)
 
     // Grab addresses:
@@ -169,9 +181,10 @@ export function makeSwapuzPlugin(opts: EdgeCorePluginOptions): EdgeSwapPlugin {
     const getQuote = async (mode: 'fix' | 'float'): Promise<SwapOrder> => {
       const { nativeAmount } = request
 
-      const largeDenomAmount = await fromWallet.nativeToDenomination(
+      const largeDenomAmount = nativeToDenomination(
+        fromWallet,
         nativeAmount,
-        fromCurrencyCode
+        request.fromTokenId
       )
 
       const getRateResponse = await fetchCors(
@@ -190,9 +203,10 @@ export function makeSwapuzPlugin(opts: EdgeCorePluginOptions): EdgeSwapPlugin {
           const requiredMatch = json.message.match(/required ([0-9.]+)/)
           if (requiredMatch?.[1] != null) {
             const minAmount = requiredMatch[1]
-            const nativeMin = await fromWallet.denominationToNative(
+            const nativeMin = denominationToNative(
+              fromWallet,
               minAmount,
-              fromCurrencyCode
+              request.fromTokenId
             )
             throw new SwapBelowLimitError(swapInfo, nativeMin)
           }
@@ -216,9 +230,10 @@ export function makeSwapuzPlugin(opts: EdgeCorePluginOptions): EdgeSwapPlugin {
       const { minAmount } = getRateJson.result
 
       if (gt(minAmount.toString(), largeDenomAmount)) {
-        const nativeMinAmount = await fromWallet.denominationToNative(
+        const nativeMinAmount = denominationToNative(
+          fromWallet,
           minAmount.toString(),
-          fromCurrencyCode
+          request.fromTokenId
         )
         throw new SwapBelowLimitError(swapInfo, nativeMinAmount)
       }
@@ -248,9 +263,10 @@ export function makeSwapuzPlugin(opts: EdgeCorePluginOptions): EdgeSwapPlugin {
         const textArray = text.split(' ')
         if (textArray[7] === 'maxAmount' && !isNaN(parseFloat(textArray[12]))) {
           const nativeMaxAmount = floor(
-            await fromWallet.denominationToNative(
+            denominationToNative(
+              fromWallet,
               textArray[12],
-              fromCurrencyCode
+              request.fromTokenId
             ),
             0
           )
@@ -279,9 +295,10 @@ export function makeSwapuzPlugin(opts: EdgeCorePluginOptions): EdgeSwapPlugin {
       } = createOrderJson.result
 
       const toNativeAmount = floor(
-        await toWallet.denominationToNative(
+        denominationToNative(
+          toWallet,
           amountResult.toString(),
-          toCurrencyCode
+          request.toTokenId
         ),
         0
       )
@@ -381,9 +398,10 @@ export function makeSwapuzPlugin(opts: EdgeCorePluginOptions): EdgeSwapPlugin {
         // Must make a copy of the request because this is a shared object
         // reused between requests to other exchange plugins
         const requestToHack = { ...requestTop }
-        const requestToExchangeAmount = await toWallet.nativeToDenomination(
+        const requestToExchangeAmount = nativeToDenomination(
+          toWallet,
           nativeAmount,
-          toCurrencyCode
+          requestTop.toTokenId
         )
         let fromQuoteNativeAmount = nativeAmount
         let retries = 5
@@ -402,9 +420,10 @@ export function makeSwapuzPlugin(opts: EdgeCorePluginOptions): EdgeSwapPlugin {
             swapOrder.spendInfo.savedAction?.toAsset.nativeAmount
           if (destNativeAmount == null) break
 
-          const toExchangeAmount = await toWallet.nativeToDenomination(
+          const toExchangeAmount = nativeToDenomination(
+            toWallet,
             destNativeAmount,
-            toCurrencyCode
+            requestTop.toTokenId
           )
           if (gte(toExchangeAmount, requestToExchangeAmount)) {
             return await makeSwapPluginQuote(swapOrder)