+++ src/swap/defi/thorchain/swapkit.ts
@@ -22,16 +22,18 @@ import {
 } from 'edge-core-js/types'
 
 import {
-  checkInvalidCodes,
+  checkInvalidTokenIds,
   getMaxSwappable,
   makeSwapPluginQuote,
   SwapOrder
 } from '../../../util/swapHelpers'
 import {
   convertRequest,
+  denominationToNative,
   fetchInfo,
   fetchWaterfall,
   getAddress,
+  nativeToDenomination,
   promiseWithTimeout
 } from '../../../util/utils'
 import { EdgeSwapRequestPlugin } from '../../types'
@@ -43,7 +45,7 @@ import {
   EXCHANGE_INFO_UPDATE_FREQ_MS,
   EXPIRATION_MS,
   getGasLimit,
-  INVALID_CURRENCY_CODES
+  INVALID_TOKEN_IDS
 } from './thorchainCommon'
 
 const pluginId = 'swapkit'
@@ -182,7 +184,7 @@ export function makeSwapKitPlugin(opts: EdgeCorePluginOptions): EdgeSwapPlugin {
     // Do not support transfer between same assets
     if (
       fromWallet.currencyInfo.pluginId === toWallet.currencyInfo.pluginId &&
-      request.fromCurrencyCode === request.toCurrencyCode
+      request.fromTokenId === request.toTokenId
     ) {
       throw new SwapCurrencyError(swapInfo, request)
     }
@@ -192,7 +194,7 @@ export function makeSwapKitPlugin(opts: EdgeCorePluginOptions): EdgeSwapPlugin {
     let daVolatilitySpread: number = DA_VOLATILITY_SPREAD_DEFAULT
     const thorswapServers: string[] = THORSWAP_DEFAULT_SERVERS
 
-    checkInvalidCodes(INVALID_CURRENCY_CODES, request, swapInfo)
+    checkInvalidTokenIds(INVALID_TOKEN_IDS, request, swapInfo)
 
     // Grab addresses:
     const fromAddress = await getAddress(fromWallet)
@@ -248,9 +250,10 @@ export function makeSwapKitPlugin(opts: EdgeCorePluginOptions): EdgeSwapPlugin {
       throw new SwapCurrencyError(swapInfo, request)
     }
 
-    const sellAmount = await fromWallet.nativeToDenomination(
+    const sellAmount = nativeToDenomination(
+      fromWallet,
       nativeAmount,
-      fromCurrencyCode
+      fromTokenId
     )
 
     const quoteParams: ThorSwapQuoteParams = {
@@ -322,7 +325,7 @@ export function makeSwapKitPlugin(opts: EdgeCorePluginOptions): EdgeSwapPlugin {
     const { expectedBuyAmount, providers, targetAddress, expiration } = thorSwap
 
     const toNativeAmount = toFixed(
-      await toWallet.denominationToNative(expectedBuyAmount, toCurrencyCode),
+      denominationToNative(toWallet, expectedBuyAmount, toTokenId),
       0,
       0
     )